network/connector/requirements.txt
paho-mqtt == 1.6.1
pandas
ipfsapi == 0.4.4
